move publishers file and update imports
