+++ frontend/src/components/ActiveUsers/ActiveUserItem.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { User } from "../../app/features/types";
-import { handleInviteToChatClick } from '../../services/publishers';
+import { handleInviteToChatClick } from '../../services/socket/publishers';
 
 interface ActiveUserItemProps {
   user: User;
